pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jwt-drishti"
version = "0.1.0"
description = "Lightweight JWT analysis and validation CLI"
authors = [
  {name = "Yash Shendge (generated)"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
  "pyjwt>=2.0.0",
  "click>=8.0",
  "rich>=10.0"
]

[project.scripts]
jwt-drishti = "jwt_drishti.cli:main"
